Use consistent name of signal folders and ids in schema (#56)
The OTel spec names the folder as `trace`, `metrics` and `logs`. We should follow the same
for our folders where the samples are located. For the website though, we name Traces (in plural).

This is in preparation for adding samples for `metrics` and `logs`. It does not affect any of the existing samples.